+++ platforms/android/src/com/mb/android/preferences/Preferences.java
@@ -1,4 +1,4 @@
-package com.mb.android.media;
+package com.mb.android.preferences;
 
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.Editor;
